### 2021-12-17 #4005 @whedon
[PRE REVIEW]: PsiRESP: calculating RESP charges with Psi4
**Submitting author:** @lilyminium (<a href="http://orcid.org/0000-0002-6095-6704">Lily Wang</a>)
**Repository:** <a href="https://github.com/lilyminium/psiresp" target ="_blank">https://github.com/lilyminium/psiresp</a>
**Version:** v0.3.1
