### 2021-12-17 ead4561
remove doi prefix
