+++ paper/paper.bib
@@ -245,7 +245,7 @@
   issn       = {1096-987X},
   shorttitle = {Fast, efficient generation of high-quality atomic charges. {AM1}-{BCC} model},
   url        = {http://www.onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291096-987X%2820000130%2921%3A2%3C132%3A%3AAID-JCC5%3E3.0.CO%3B2-P},
-  doi        = {https://doi.org/10.1002/(SICI)1096-987X(20000130)21:2<132::AID-JCC5>3.0.CO;2-P},
+  doi        = {10.1002/(SICI)1096-987X(20000130)21:2<132::AID-JCC5>3.0.CO;2-P},
   abstract   = {The AM1-BCC method quickly and efficiently generates high-quality atomic charges for use in condensed-phase simulations. The underlying features of the electron distribution including formal charge and delocalization are first captured by AM1 atomic charges for the individual molecule. Bond charge corrections (BCCs), which have been parameterized against the HF/6-31G* electrostatic potential (ESP) of a training set of compounds containing relevant functional groups, are then added using a formalism identical to the consensus BCI (bond charge increment) approach. As a proof of the concept, we fit BCCs simultaneously to 45 compounds including O-, N-, and S-containing functionalities, aromatics, and heteroaromatics, using only 41 BCC parameters. AM1-BCC yields charge sets of comparable quality to HF/6-31G* ESP-derived charges in a fraction of the time while reducing instabilities in the atomic charges compared to direct ESP-fit methods. We then apply the BCC parameters to a small “test set” consisting of aspirin, d-glucose, and eryodictyol; the AM1-BCC model again provides atomic charges of quality comparable with HF/6-31G* RESP charges, as judged by an increase of only 0.01 to 0.02 atomic units in the root-mean-square (RMS) error in ESP. Based on these encouraging results, we intend to parameterize the AM1-BCC model to provide a consistent charge model for any organic or biological molecule. © 2000 John Wiley \& Sons, Inc. J Comput Chem 21: 132–146, 2000},
   language   = {en},
   number     = {2},
@@ -265,7 +265,7 @@
   copyright = {Copyright © 1993 John Wiley \& Sons, Inc.},
   issn      = {1096-987X},
   url       = {http://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.540141013},
-  doi       = {https://doi.org/10.1002/jcc.540141013},
+  doi       = {10.1002/jcc.540141013},
   abstract  = {Absolute free energies of hydration have been computed for 13 diverse organic molecules using partial charges derived from ab initio 6-31G* wave functions. Both Mulliken charges and charges fit to the electrostatic potential surface (EPS) were considered in conjunction with OPLS Lennard–Jones parameters for the organic molecules and the TIP4P model of water. Monte Carlo simulations with statistical perturbation theory yielded relative free energies of hydration. These were converted to absolute quantities through perturbations to reference molecules for which absolute free energies of hydration had been obtained previously in TIP4P water. The average errors in the computed absolute free energies of hydration are 1.1 kcal/mol for the 6-31G* EPS charges and 4.0 kcal/mol for the Mulliken charges. For the EPS charges, the largest individual errors are under 2 kcal/mol except for acetamide, in which case the error is 3.7 kcal/mol. The hydrogen bonding between the organic solutes and water has also been characterized. © John Wiley \& Sons, Inc.},
   language  = {en},
   number    = {10},
@@ -354,7 +354,7 @@
   number   = {2},
   pages    = {e1491},
[…]
