### 2021-12-18 e1ddcea
add more detail to docs
