### 2021-12-18 9c23fce
rename configured classes
