+++ docs/source/_autosummary/psiresp.moleculebase.BaseMolecule.rst
@@ -1,6 +1,0 @@
-﻿psiresp.moleculebase.BaseMolecule
-=================================
-
-.. currentmodule:: psiresp.moleculebase
-
-.. autopydantic_model:: BaseMolecule
