+++ docs/source/_autosummary/psiresp.configs.RespA1.rst
@@ -1,6 +1,0 @@
-﻿psiresp.configs.RespA1
-======================
-
-.. currentmodule:: psiresp.configs
-
-.. autopydantic_model:: RespA1
