+++ docs/source/_autosummary/psiresp.resp.RespOptions.rst
@@ -1,6 +1,0 @@
-﻿psiresp.resp.RespOptions
-========================
-
-.. currentmodule:: psiresp.resp
-
-.. autopydantic_model:: RespOptions
