+++ docs/source/_autosummary/psiresp.conformer.ConformerGenerationOptions.rst
@@ -1,6 +1,0 @@
-﻿psiresp.conformer.ConformerGenerationOptions
-============================================
-
-.. currentmodule:: psiresp.conformer
-
-.. autopydantic_model:: ConformerGenerationOptions

[…]
