### 2021-12-18 e15d702
switch to static plot
