### 2021-12-20 0c4dc66
Add more detail to documentation (#58)
