+++ docs/source/examples/README.rst
@@ -1,3 +1,4 @@
+.. _examples-label:
 
 ========
 Examples
