### 2022-01-15 982ad93
reshuffle to avoid rdkit, psi4 and qcfractal
