+++ devtools/conda-envs/environment.yaml
@@ -15,6 +15,7 @@ dependencies:
   - psi4 >= 1.3.0
   - intel-openmp != 2019.5
   - geometric
+  - msgpack-python
 
   # qcstack
   - postgresql >= 13.0
