+++ devtools/conda-envs/minimal_test_env.yaml
@@ -15,6 +15,7 @@ dependencies:
   # deps
   - numpy
   - scipy >= 1.4
+  - tqdm
 
   # Pip-only installs
   - pip:
