+++ devtools/conda-envs/test_env.yaml
@@ -15,6 +15,7 @@ dependencies:
   - psi4 >= 1.3.0
   - intel-openmp!=2019.5
   - geometric
+  - msgpack-python
 
   # qcstack
   - postgresql >= 13.0
@@ -28,6 +29,7 @@ dependencies:
   - rdkit
   - numpy
   - scipy >= 1.4
+  - tqdm
 
   # weird mamba mismatch
   - h5py<3.2
@@ -39,4 +41,3 @@ dependencies:
       - pytest-xdist
       - duecredit
       - importlib_resources
-      - mdanalysis>=2.0
