+++ psiresp/orientation.py
@@ -1,4 +1,3 @@
-import re
 from typing import Optional
 
 import numpy as np
