+++ devtools/conda-envs/test_env.yaml
@@ -18,7 +18,7 @@ dependencies:
 
   # qcstack
   - postgresql >= 13.0
-  - qcelemental == 0.22.0
+  - qcelemental
   - qcengine
   - qcfractal
   - qcfractal-core
