### 2022-01-16 f16192c
redo attrs and format a bit
