+++ psiresp/configs.py
@@ -50,9 +50,9 @@ class TwoStageRESP(ConfiguredJob):
                         basis="6-31g*"
                         ),
         grid_options=dict(use_radii="msk"),
-        resp_options=dict(restraint_scale_stage_1=0.0005,
-                          restraint_scale_stage_2=0.001,
-                          restraint_steepness=0.1,
+        resp_options=dict(restraint_height_stage_1=0.0005,
+                          restraint_height_stage_2=0.001,
+                          restraint_slope=0.1,
                           stage_2=True,
                           exclude_hydrogens=True,
                           restrained_fit=True),
@@ -73,9 +73,9 @@ class OneStageRESP(ConfiguredJob):
                                           basis="6-31g*"
                                           ),
                           grid_options=dict(use_radii="msk"),
-                          resp_options=dict(restraint_scale_stage_1=0.01,
-                                            restraint_scale_stage_2=0.0,
-                                            restraint_steepness=0.1,
+                          resp_options=dict(restraint_height_stage_1=0.01,
+                                            restraint_height_stage_2=0.0,
+                                            restraint_slope=0.1,
                                             stage_2=False,
                                             exclude_hydrogens=True,
                                             restrained_fit=True),
@@ -96,9 +96,9 @@ class ESP(ConfiguredJob):
                                           basis="6-31g*"
                                           ),
                           grid_options=dict(use_radii="msk"),
-                          resp_options=dict(restraint_scale_stage_1=0.0,
-                                            restraint_scale_stage_2=0.0,
-                                            restraint_steepness=0.1,
+                          resp_options=dict(restraint_height_stage_1=0.0,
+                                            restraint_height_stage_2=0.0,
+                                            restraint_slope=0.1,
                                             stage_2=False,
                                             exclude_hydrogens=True,
                                             restrained_fit=False),
@@ -123,9 +123,9 @@ class WeinerESP(ConfiguredJob):
                                           basis="sto-3g"
                                           ),
                           grid_options=dict(use_radii="msk"),
-                          resp_options=dict(restraint_scale_stage_1=0.0,
-                                            restraint_scale_stage_2=0.0,
-                                            restraint_steepness=0.1,
+                          resp_options=dict(restraint_height_stage_1=0.0,
+                                            restraint_height_stage_2=0.0,
+                                            restraint_slope=0.1,
                                             stage_2=False,
                                             exclude_hydrogens=True,
                                             restrained_fit=False),
@@ -145,9 +145,9 @@ class ATBRESP(ConfiguredJob):
                                               solvent="water"
                                           )),
                           grid_options=dict(use_radii="msk"),
-                          resp_options=dict(restraint_scale_stage_1=0.0,
-                                            restraint_scale_stage_2=0.0,
-                                            restraint_steepness=0.1,
+                          resp_options=dict(restraint_height_stage_1=0.0,
+                                            restraint_height_stage_2=0.0,
+                                            restraint_slope=0.1,
                                             stage_2=False,
                                             exclude_hydrogens=False,
                                             restrained_fit=False),
