+++ psiresp/molecule.py
@@ -1,3 +1,4 @@
+from random import random
 from typing import List, Optional, Tuple
 import functools
 import logging
@@ -102,13 +103,20 @@ class Molecule(BaseMolecule):
 
     @classmethod
     def from_rdkit(cls, molecule, random_seed=-1, **kwargs):
-        from . import rdutils
-        qcmol = rdutils.rdmol_to_qcelemental(molecule, random_seed=random_seed)
-        kwargs = dict(**kwargs)
-        kwargs["qcmol"] = qcmol
-        obj = cls(**kwargs)
-        obj._rdmol = molecule
-        return obj
+        from .rdutils import molecule_from_rdkit
+        return molecule_from_rdkit(molecule, cls, random_seed=random_seed,
+                                   **kwargs)
+
+    @property
+    def charges(self):
+        for charge_prop in [
+            self.stage_2_restrained_charges,
+            self.stage_2_unrestrained_charges,
+            self.stage_1_restrained_charges,
+            self.stage_1_unrestrained_charges,
+        ]:
+            if charge_prop is not None:
+                return charge_prop
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -120,6 +128,14 @@ class Molecule(BaseMolecule):
             self.multiplicity = self.qcmol.molecular_multiplicity
         else:
             self.qcmol.__dict__["molecular_multiplicity"] = self.multiplicity
+
+    def to_rdkit(self):
+        from .rdutils import molecule_to_rdkit
+        return molecule_to_rdkit(self)
+
+    def to_mdanalysis(self):
+        from .mdautils import molecule_to_mdanalysis
+        return molecule_to_mdanalysis(self)
 
     def __repr__(self):
         qcmol_repr = self._get_qcmol_repr()
