+++ .github/workflows/gh-ci.yaml
@@ -25,23 +25,23 @@ jobs:
         codecov: [true]
         runslow: [false]
         environment: [test_env]
-        exclude:
-          # Tests hang. Not sure why
-          - os: ubuntu-latest
-            python-version: 3.7
-          - os: ubuntu-latest
-            python-version: 3.9
+        # exclude:
+          # # Tests hang. Not sure why
+          # - os: ubuntu-latest
+          #   python-version: 3.8
+          # - os: ubuntu-latest
+          #   python-version: 3.9
         include:
           - os: ubuntu-latest
-            python-version: 3.9
+            python-version: 3.7
             runslow: true
             environment: test_env
           - os: macOS-latest
-            python-version: 3.7
+            python-version: 3.8
             runslow: true
             environment: test_env
           - os: ubuntu-latest
-            python-version: 3.8
+            python-version: 3.9
             environment: minimal_test_env
             runslow: false
 
