+++ psiresp/molecule.py
@@ -102,20 +102,13 @@ class Molecule(BaseMolecule):
 
     @classmethod
     def from_rdkit(cls, molecule, random_seed=-1, **kwargs):
-        from .rdutils import molecule_from_rdkit
-        return molecule_from_rdkit(molecule, cls, random_seed=random_seed,
-                                   **kwargs)
-
-    @property
-    def charges(self):
-        for charge_prop in [
-            self.stage_2_restrained_charges,
-            self.stage_2_unrestrained_charges,
-            self.stage_1_restrained_charges,
-            self.stage_1_unrestrained_charges,
-        ]:
-            if charge_prop is not None:
-                return charge_prop
+        from . import rdutils
+        qcmol = rdutils.rdmol_to_qcelemental(molecule, random_seed=random_seed)
+        kwargs = dict(**kwargs)
+        kwargs["qcmol"] = qcmol
+        obj = cls(**kwargs)
+        obj._rdmol = molecule
+        return obj
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -127,14 +120,6 @@ class Molecule(BaseMolecule):
             self.multiplicity = self.qcmol.molecular_multiplicity
         else:
             self.qcmol.__dict__["molecular_multiplicity"] = self.multiplicity
-
-    def to_rdkit(self):
-        from .rdutils import molecule_to_rdkit
-        return molecule_to_rdkit(self)
-
-    def to_mdanalysis(self):
-        from .mdautils import molecule_to_mdanalysis
-        return molecule_to_mdanalysis(self)
 
     def __repr__(self):
         qcmol_repr = self._get_qcmol_repr()
