+++ .github/workflows/gh-ci.yaml
@@ -25,15 +25,15 @@ jobs:
         codecov: [true]
         runslow: [false]
         environment: [test_env]
-        # exclude:
-          # # Tests hang. Not sure why
-          # - os: ubuntu-latest
-          #   python-version: 3.8
+        exclude:
+          # Tests hang. Not sure why
+          - os: ubuntu-latest
+            python-version: 3.7
           # - os: ubuntu-latest
           #   python-version: 3.9
         include:
           - os: ubuntu-latest
-            python-version: 3.7
+            python-version: 3.9
             runslow: true
             environment: test_env
           - os: macOS-latest
@@ -41,7 +41,7 @@ jobs:
             runslow: true
             environment: test_env
           - os: ubuntu-latest
-            python-version: 3.9
+            python-version: 3.7
             environment: minimal_test_env
             runslow: false
 
