### 2022-01-16 0bca066
Rename configured classes and RespOptions variable (#59)
Fixes #56
