### 2022-01-16 53130fa
Rearrange code to avoid RDKit, Psi4 and QCFractal where possible (#60)
-  rewrites some functions in pure Python where possible
- wraps imports in functions
- adds a minimal suite of tests
