+++ setup.py
@@ -19,7 +19,6 @@ except FileNotFoundError:
     long_description = "\n".join(short_description[2:])
 
 setup(
-    # Self-descriptive entries which should always be present
     name='psiresp',
     author='Lily Wang',
     author_email='lily.wang@anu.edu.au',
@@ -29,27 +28,19 @@ setup(
     version=versioneer.get_version(),
     cmdclass=versioneer.get_cmdclass(),
     license='LGPLv3',
-
-    # Which Python importable modules should be included when your package is installed
-    # Handled automatically by setuptools. Use 'exclude' to prevent some specific
-    # subpackage(s) from being added, if needed
     packages=find_packages(),
-
-    # Optional include package data to ship with your package
-    # Customize MANIFEST.in if the general case does not suit your needs
-    # Comment out this line to prevent the files from being packaged with your software
     include_package_data=True,
-
-    # Allows `setup.py test` to work correctly with pytest
     setup_requires=[] + pytest_runner,
+    install_requires=[
+        'msgpack-python',
+        'numpy',
+        'qcelemental',
+        'requests',
+        'scipy>=1.4',
+        'tqdm',
+        'typing-extensions',
 
-    # Additional entries you may want simply uncomment the lines you want and fill in the data
-    # url='http://www.my_package.com',  # Website
-    install_requires=['numpy', 'qcelemental',
-                      'duecredit'],  # Required packages, pulls from pip if needed; do not use for Conda deployment
-    platforms=['Linux', 'Mac OS-X', 'Unix'],  # Valid platforms your code works on, adjust to your flavor
-    python_requires=">=3.7",  # Python version restrictions
-
-    # Manual control if final package is compressible or not, set False to prevent the .egg from being made
-    # zip_safe=False,
+    ],
+    platforms=['Linux', 'Mac OS-X', 'Unix'],
+    python_requires=">=3.7",
 )
