### 2022-01-17 #4005 @lilyminium
@pdebuyl I just made a new release of psiresp 0.3 -- would it be possible to change the version to that instead? It contains minor improvements:  making some dependencies optional and renaming some ambiguously-named variables.
