### 2022-01-26 d125c54
add notes about conda install
