### 2022-01-26 db44705
remove extra brackets
