+++ README.md
@@ -1,9 +1,9 @@
 # PsiRESP
 😪-RESP
 
-| **Latest release** | [[![Last release tag](https://img.shields.io/github/release-pre/lilyminium/psiresp.svg)](https://github.com/lilyminium/psiresp/releases) ![GitHub commits since latest release (by date) for a branch](https://img.shields.io/github/commits-since/lilyminium/psiresp/latest)  [![Documentation Status](https://readthedocs.org/projects/psiresp/badge/?version=latest)](https://psiresp.readthedocs.io/en/latest/?badge=latest)|
+| **Latest release** | [![Last release tag](https://img.shields.io/github/release-pre/lilyminium/psiresp.svg)](https://github.com/lilyminium/psiresp/releases) ![GitHub commits since latest release (by date) for a branch](https://img.shields.io/github/commits-since/lilyminium/psiresp/latest)  [![Documentation Status](https://readthedocs.org/projects/psiresp/badge/?version=latest)](https://psiresp.readthedocs.io/en/latest/?badge=latest)|
 | :------ | :------- |
-| **Installation** | [![Conda Recipe](https://img.shields.io/badge/recipe-psiresp-green.svg)](https://anaconda.org/conda-forge/psiresp) ![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/psiresp.svg)] [![PyPI version](https://badge.fury.io/py/psiresp.svg)](https://badge.fury.io/py/psiresp) ![PyPI - Downloads](https://img.shields.io/pypi/dm/psiresp) |
+| **Installation** | [![Conda Recipe](https://img.shields.io/badge/recipe-psiresp-green.svg)](https://anaconda.org/conda-forge/psiresp) ![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/psiresp.svg) [![PyPI version](https://badge.fury.io/py/psiresp.svg)](https://badge.fury.io/py/psiresp) ![PyPI - Downloads](https://img.shields.io/pypi/dm/psiresp) |
 | **Status** | [![GH Actions Status](https://github.com/lilyminium/psiresp/actions/workflows/gh-ci.yaml/badge.svg)](https://github.com/lilyminium/psiresp/actions?query=branch%3Amaster+workflow%3Agh-ci) [![codecov](https://codecov.io/gh/lilyminium/psiresp/branch/master/graph/badge.svg)](https://codecov.io/gh/lilyminium/psiresp/branch/master) [![Language grade: Python](https://img.shields.io/lgtm/grade/python/g/lilyminium/psiresp.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/lilyminium/psiresp/context:python) |
 | **Community** | [![License: LGPL v3](https://img.shields.io/badge/License-LGPL_v3-blue.svg)](https://www.gnu.org/licenses/lgpl-3.0) |
 
