+++ docs/source/installation.rst
@@ -2,8 +2,8 @@ Installation
 ============
 
 Much of the functionality in PsiRESP depends on `RDKit`_, `Psi4`_,
-and `QCFractal`_ . These packages are easiest distributed through
-`Anaconda`_ , so it is **highly recommended** to install PsiRESP
+and `QCFractal`_. These packages are easiest distributed through
+`Anaconda`_, so it is **highly recommended** to install PsiRESP
 through ``conda``.
 
 .. code-block:: bash
