### 2022-01-26 74841e7
Add notes about conda install (#61)
