+++ .github/workflows/gh-ci.yaml
@@ -17,6 +17,7 @@ jobs:
   main_tests:
     if: "github.repository == 'lilyminium/psiresp'"
     runs-on: ${{ matrix.os }}
+    timeout-minutes: 60
     strategy:
       fail-fast: false
       matrix:
