### 2022-01-26 e755a1b
Add notes about conda install (#61)
