+++ devtools/conda-envs/test_env.yaml
@@ -18,8 +18,8 @@ dependencies:
 
   # qcstack
   - postgresql >= 13.0
-  - qcelemental == 0.22.0
-  - qcengine
+  - qcelemental == 0.24.0
+  - qcengine == 0.21
   - qcfractal
   - qcfractal-core
   - requests
