+++ .github/workflows/gh-ci.yaml
@@ -96,13 +96,13 @@ jobs:
         if: ${{ matrix.runslow == false }}
         run: |
           PYTEST_FLAGS=" --disable-pytest-warnings --durations=50 --cov=psiresp --cov-report=xml"
-          pytest psiresp/tests $PYTEST_FLAGS
+          pytest -n 4 psiresp/tests $PYTEST_FLAGS
 
       - name: run_slow_tests
         if: ${{ matrix.runslow == true }}
         run: |
           PYTEST_FLAGS=" --disable-pytest-warnings --durations=50 --cov=psiresp --cov-report=xml"
-          pytest --runslow psiresp/tests $PYTEST_FLAGS
+          pytest -n 4 --runslow psiresp/tests $PYTEST_FLAGS
         
       - name: codecov
         uses: codecov/codecov-action@v1
