### 2022-01-26 #4005 @pdebuyl
@hannahbrucemacdonald @ptmerz would you review the article "PsiRESP: calculating RESP charges with Psi4" for the Journal of Open Source Software?
