+++ psiresp/tests/conftest.py
@@ -14,30 +14,27 @@ pytest_plugins = [
 ]
 
 
-# @pytest.fixture(scope="session")
-# def postgres_server():
-#     storage = TemporaryPostgres(database_name="test_psiresp")
-#     storage.psql.restore_database(POSTGRES_SERVER_BACKUP)
-#     return storage.psql
-#     storage.stop()
+@pytest.fixture(scope="session")
+def postgres_server():
+    pytest.importorskip("qcfractal.postgres_harness")
+    from psiresp.testing import TemporaryPostgres
+    storage = TemporaryPostgres(database_name="test_psiresp")
+    storage.psql.restore_database(POSTGRES_SERVER_BACKUP)
+    yield storage.psql
+    storage.stop()
 
 
 @pytest.fixture(scope="session")
-def fractal_client():
-    pytest.importorskip("polytop")
+def fractal_client(postgres_server):
     pytest.importorskip("qcfractal.postgres_harness")
     from qcfractal import FractalSnowflake, interface as ptl
-    from psiresp.testing import TemporaryPostgres
 
-    storage = TemporaryPostgres(database_name="test_psiresp")
-    storage.psql.restore_database(POSTGRES_SERVER_BACKUP)
-    postgres_server = storage.psql
     with FractalSnowflake(
-        max_workers=1,
+        # max_workers=1,
         storage_project_name="test_psiresp",
         storage_uri=postgres_server.database_uri(),
         reset_database=False,
-        start_server=False,
+        # start_server=False,
     ) as server:
         yield ptl.FractalClient(server)
 
