### 2022-01-26 #4100 @whedon
[REVIEW]: PsiRESP: calculating RESP charges with Psi4
**Submitting author:** @lilyminium (<a href="http://orcid.org/0000-0002-6095-6704">Lily Wang</a>)
**Repository:** https://github.com/lilyminium/psiresp
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.2
