### 2022-01-26 #4100 @pdebuyl
@ptmerz @hannahbrucemacdonald thanks for accepting to review for JOSS! Make sure to accept the invitation to the reviewers group and to have a look at the reviewer guidelines linked to at the top of this review page.

The review process will happen in this issue page, so questions to the author or to me can be added as comments here.
