+++ .github/workflows/gh-ci.yaml
@@ -35,7 +35,7 @@ jobs:
             environment: test_env
             runslow: true
           - os: ubuntu-latest
-            python-version: 3.9
+            python-version: 3.10
             environment: minimal_test_env
             runslow: true
 
