+++ .github/workflows/gh-ci.yaml
@@ -28,12 +28,12 @@ jobs:
         include:
           - os: macOS-latest
             python-version: 3.7
-            runslow: true
+            runslow: false
             environment: test_env
           - os: ubuntu-latest
             python-version: 3.7
             environment: test_env
-            runslow: true
+            runslow: false
           - os: ubuntu-latest
             python-version: "3.10"
             environment: minimal_test_env
