+++ psiresp/tests/conftest.py
@@ -16,7 +16,7 @@ pytest_plugins = [
 
 
 @requires_qcfractal
-# @pytest.fixture(scope="session")
+@pytest.fixture(scope="session")
 def postgres_server():
     pytest.importorskip("qcfractal.postgres_harness")
     from psiresp.testing import TemporaryPostgres
@@ -27,8 +27,9 @@ def postgres_server():
 
 
 @requires_qcfractal
-# @pytest.fixture(scope="session")
+@pytest.fixture(scope="session")
 def fractal_client(postgres_server):
+    pytest.importorskip("polytop")
     pytest.importorskip("qcfractal.postgres_harness")
     from qcfractal import FractalSnowflake, interface as ptl
 
