### 2022-01-27 2c5ac3a
pin qcel to version used to generate qcfractal
