+++ psiresp/tests/test_qm.py
@@ -25,7 +25,7 @@ class TestQMEnergyOptions:
         assert record.status == "INCOMPLETE"
         assert record.get_molecule() == dmso_qcmol
 
-    @pytest.mark.slow
+    # @pytest.mark.slow
     def test_add_compute_and_wait(self, cheap_options, dmso_qcmol, empty_client):
         molhash = dmso_qcmol.get_hash()
         records = empty_client.query_molecules(molecule_hash=[molhash])
