### 2022-01-27 14f6922
put qcfractal installs back in
