### 2022-01-27 b032407
try empty?
