### 2022-01-27 a4f72fd
comment out qm tests
