+++ .github/workflows/gh-ci.yaml
@@ -20,16 +20,16 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macOS-latest]
+        os: [macOS-latest]
         python-version: [3.7, 3.8, 3.9]
         codecov: [true]
         runslow: [true]
         environment: [no_qcfractal_test_env]
-        exclude:
-          - os: ubuntu-latest
-            python-version: 3.7
-          - os: ubuntu-latest
-            python-version: 3.8
+        # exclude:
+        #   - os: ubuntu-latest
+        #     python-version: 3.7
+        #   - os: ubuntu-latest
+        #     python-version: 3.8
         include:
           - os: ubuntu-latest
             python-version: 3.7
