### 2022-01-28 dda44bd
fix np empty
