+++ .github/workflows/gh-ci.yaml
@@ -21,13 +21,10 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macOS-latest]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.8, 3.9]
         codecov: [true]
         runslow: [true]
         environment: [no_qcfractal_test_env]
-        exclude:
-          - os: ubuntu-latest
-            python-version: 3.7
         include:
           - os: macOS-latest
             python-version: 3.7
