+++ .github/workflows/gh-ci.yaml
@@ -28,6 +28,8 @@ jobs:
         exclude:
           - os: ubuntu-latest
             python-version: 3.7
+          - os: ubuntu-latest
+            python-version: 3.9
         include:
           - os: ubuntu-latest
             python-version: 3.7
