### 2022-01-28 c292a05
Default to tests running without QCFractal (#62)
Fixes #63
