+++ psiresp/tests/utils.py
@@ -1,5 +1,16 @@
+import pytest
+
 import numpy as np
 import qcelemental as qcel
+
+try:
+    import qcfractal.interface
+except ImportError:
+    qcfractal_is_installed = False
+else:
+    qcfractal_is_installed = True
+
+requires_qcfractal = pytest.mark.skipif(not qcfractal_is_installed, reason="requires QCFractal")
 
 
 def load_gamess_esp(file):
