### 2022-01-28 c85b6c6
Default to tests running without QCFractal (#62)
Fixes #63
