### 2022-01-28 c494729
change solver to least squares
