### 2022-01-28 1928cce
change coverage settings
