### 2022-01-28 5d73550
Change coverage settings (#66)
