### 2022-01-28 36c2ba4
check with gssv before solving
