### 2022-01-28 c61ac08
Add explicit check for singular matrices (#65)
