+++ .github/workflows/gh-ci.yaml
@@ -32,8 +32,6 @@ jobs:
         exclude:
           - os: ubuntu-latest
             python-version: 3.7
-          - os: ubuntu-latest
-            python-version: 3.9
         include:
           - os: ubuntu-latest
             python-version: 3.7
