### 2022-01-28 #4100 @ptmerz
Hi @lilyminium, I started the review today. I didn't have time yet to review the package in all details, so I won't comment on it yet, but it definitely looks to be in great shape!

I did review the paper already, and I think it's very well written. A few things I noted:
* The author instructions (https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) ask to start the paper with "a summary of the high-level functionality of your software for a non-specialist reader". This section is missing in your paper.
* I think the Statement of Need section could be split into two sections. Most of this section (from the second paragraph on) actually describes the functionality of the software and isn't a statement of need, so this could be split off in a new section titled "Functionality" or something similar.
* I think you could extend the Statement of Need section to include some real-world use cases. You are doing a great job at explaining the existing packages, and how your package improves upon them, so it would be interesting to add some real-world examples of what these improvements can achieve (or make easier) compared to the existing solutions.
* You are repeatedly adding citations for the same concepts or methods in the different sections. I think that adding citations at the first mentioning is enough and could improve the readability of the remaining text, but maybe @pdebuyl can chime in if there are some editorial recommendations regarding this.
