+++ paper/paper.md
@@ -20,11 +20,17 @@ date: 16 December 2021
 bibliography: paper.bib
 ---
 
+
+
+# Summary
+
+Molecular dynamics (MD) simulations study the movement of particles over time, and have become a fundamental tool in biomedical and materials research. The accuracy of a MD simulation depends on its ability to model the physics of the chosen system, a capacity that arises from the parameters used to model the interactions between atoms (also known as a “force field”). The functional form of most force fields is an approximation of real word physics to allow for tractable simulation; one common simplification is to model the dynamic distribution of electrons around an atom as single, static, fixed partial charge. Electrostatic interactions between non-bonded atoms can then be simplified to pairwise interactions between the partial charges.
+
+PsiRESP is a Python package that uses the Psi4 quantum chemistry engine to calculate atomic partial charges. It supports multiple methods, each based on the electrostatic potential experienced at particular grid points around the molecule. These partial charges can then be used in MD simulations.
+
 # Background
 
-Molecular dynamics (MD) simulations have become a fundamental tool in biomedical and materials research. The accuracy of a MD simulation depends on its ability to model the physics of the chosen system, a capacity that arises from the parameters used to model bonded and non-bonded interactions between atoms (also known as a “force field”). Non-bonded interactions comprise Lennard-Jones and electrostatic interactions. Electrostatic interactions between non-bonded atoms are typically modelled as pairwise interactions between atomic partial charges. 
-
-A number of methods have been developed to generate molecular partial charges, and results vary widely between each method. Approaches based on quantum mechanics calculations typically follow one of three strategies: wavefunction population analysis (e.g. Mulliken charges [@mulliken1955], Löwdin charges [@lowdin1950]); partitioning the distribution of electron density (e.g. Hirshfield charges [@hirshfeld1977]); or reproducing the electrostatic potential (ESP) outside the molecular surface. The major classical biomolecular force fields generally obtain charges using the last strategy.
+A number of methods have been developed to generate molecular partial charges, and results vary widely between each method. Approaches based on quantum mechanics (QM) calculations are particularly advantageous for novel compounds, as no knowledge beyond the molecular structure is required. One common approach is to reproduce the electrostatic potential (ESP) outside the molecular surface.
 
 ## ESP
 
@@ -55,17 +61,27 @@ Finally, a Python RESP plugin for Psi4 also exists [@alenaizan2020]. Also called
 
 # Statement of Need
 
-PsiRESP is a Python package that can be used to calculate ESP [@singh1984] and RESP charges [@bayly1993; @cornell1993; @cieplak1995], as well as the next-generation RESP2 scheme [@schauperl2020]. Both intra-molecular and inter-molecular charge constraints and charge equivalence constraints are supported. Users may generate their own conformers of a molecule, or conformers can be automatically generated. Multiple orientations can also be specified or automatically generated for each conformer, to ensure a generally applicable distribution of charges. Multiple atomic radii sets are supported, including the Bondi and Merz-Singh-Kollman sets of radii.
+PsiRESP is a Python package that can be used to calculate ESP and RESP charges, as well as the next-generation RESP2 scheme. Both intra-molecular and inter-molecular charge constraints and charge equivalence constraints are supported. Users may generate their own conformers of a molecule, or conformers can be automatically generated. Multiple orientations can also be specified or automatically generated for each conformer, to ensure a generally applicable distribution of charges. Multiple atomic radii sets are supported, including the Bondi and Merz-Singh-Kollman sets of radii.
 
[…]
