### 2022-01-28 #4100 @lilyminium
Thanks for your feedback, @ptmerz! I have opened a PR updating the paper at https://github.com/lilyminium/psiresp/pull/68 -- does that  address your comments?
