+++ paper/paper.md
@@ -53,7 +53,7 @@ The HF/6-31G* level of theory commonly used for RESP charge derivation overestim
 
 ## Existing tools
 
-Several tools already exist for fitting ESP-based charges. The canonical program is `resp` [@bayly1993],written in FORTRAN and distributed by the AmberTools package [@case]. However, using this with multiple molecules is a tedious and manual process. As the charge distribution of a molecule is highly dependent on its conformation, it is often necessary to examine multiple conformers and orientations of a molecule to obtain a general charge distribution that can be used to derive atomic partial charges [@reynolds1992]. In order to use the `resp` program, different conformers and orientations must be manually generated, individual QM jobs generated for each, and then data converted into the RESP format. Constraints are manually specified on multiple rows, which requires significant effort to input manually and is difficult to read and verify.
+Several tools already exist for fitting ESP-based charges. The canonical program is `resp` [@bayly1993], written in FORTRAN and distributed by the AmberTools package [@case]. However, using this with multiple molecules is a tedious and manual process. As the charge distribution of a molecule is highly dependent on its conformation, it is often necessary to examine multiple conformers and orientations of a molecule to obtain a general charge distribution that can be used to derive atomic partial charges [@reynolds1992]. In order to use the `resp` program, different conformers and orientations must be manually generated, individual QM jobs generated for each, and then data converted into the RESP format. Constraints are manually specified on multiple rows, which requires significant effort to input manually and is difficult to read and verify.
 
 The online R.E.D. Tools (RESP and ESP charge Derive) [@dupradeau2007; @dupradeau2010] provide a much more accessible interface to the `resp` tool. It supports quantum chemistry jobs using both the Gaussian [@gaussian] and GAMESS [@gamess] QM engines, as well as both CHELPG [@breneman1990] and Connolly [@connolly1983] surfaces. Users can furthermore specify intramolecular and intermolecular charge constraints and charge equivalence constraints. The R.E.D. Tools use the canonical ``resp`` program for the actual charge fitting. However, conformer geometries must be provided by the user, and re-orientations of each conformer manually specified. Molecules must be provided in a specialized P2N format, and it is strongly recommended that a human user check each file before submission. Moreover, while standalone tools can be downloaded and run, full functionality is only available for jobs submitted to the server.
 
