### 2022-02-02 2a86571
Update paper per review (#68)
