### 2022-02-09 lilyminium/psiresp#69 @hannahbrucemacdonald
Review for JOSS 
Linking to main thread:  https://github.com/openjournals/joss-reviews/issues/4100

Hey! I'm just getting started, but thought I'd start here.

These are my comments on the Summary. They aren't required for submission, but just suggestions.

Background - "One common approach is to reproduce the ESP outside the molecular surface.", might read better as "One common approach is to find the per-atom electrostatic charges which reproduces the ESP outside the molecular surface."

ESP - It's worth defining even the obvious parameters in the text
RESP2 - what is the higher level of QM theory that's used?
