### 2022-02-09 lilyminium/psiresp#69 @hannahbrucemacdonald
Statement of need

I think the example of ncAA is great, and really illustrates the issue. Maybe I'm not sure what the inter-molecular constraints would be though. If you had a bunch of peptides that all contain the same ncAA for example? Or something else?
