### 2022-02-09 lilyminium/psiresp#69 @hannahbrucemacdonald
For Functionality section, it says a conda package is forthcoming, but it's currently conda installable, so is the text out of date here?
