### 2022-02-09 lilyminium/psiresp#69 @hannahbrucemacdonald
I like the examples, and in the documentation and it's super super minor but it would be nice to have an example folder with the example scripts in, so that they can be run out of the box. Then if you add more documentation/features in future you can just add more files here, rather than the readme getting longer.

Also - are both examples demonstrating how to run with QCFractal? It might be helpful to have an example of how to run it locally (if I've understood the manuscript and the examples correctly)
