### 2022-02-09 lilyminium/psiresp#69 @hannahbrucemacdonald
Final comment of today is that JOSS asks for:

I might have missed it, but if you add something that aligns with what you'd like from contributors/users to the README I think that would be great!
