+++ paper/paper.md
@@ -30,7 +30,7 @@ PsiRESP is a Python package that uses the Psi4 quantum chemistry engine to calcu
 
 # Background
 
-A number of methods have been developed to generate molecular partial charges, and results vary widely between each method. Approaches based on quantum mechanics (QM) calculations are particularly advantageous for novel compounds, as no knowledge beyond the molecular structure is required. One common approach is to reproduce the electrostatic potential (ESP) outside the molecular surface.
+A number of methods have been developed to generate molecular partial charges, and results vary widely between each method. Approaches based on quantum mechanics (QM) calculations are particularly advantageous for novel compounds, as no knowledge beyond the molecular structure is required. One common approach is to find per-atom charges that reproduce the electrostatic potential outside the molecular surface. Methods following this design include the simply named "ESP" model, the restrained ESP ("RESP") model, and the "RESP2" model.
 
 ## ESP
 
@@ -65,9 +65,9 @@ PsiRESP is a Python package that can be used to calculate ESP and RESP charges, 
 
 The automated conformer generation enables computing more transferrable charges with fewer resources. Conformers are generated following the Electrostatically Least-interacting Functional groups method (ELF). Here, a pool of conformers is ranked by the electrostatic interaction between functional groups; the least-interacting conformers are selected; and then a user-specified number of final conformers is chosen from the secondary pool to maximise structural diversity (by the root mean square deviation of coordinates). This allows even a small number of conformers to cover a breadth of structural range.
 
-Intra-molecular charge constraints allow for equivalent charges between symmetric groups, as well as computing charges for a molecule that is compatible with an existing environment. For example, when calculating the charges of a non-canonical amino acid (ncAA) in a protein where the charges of canonical residues are already known, it is general practice to constrain the charges on the backbone of the ncAA to equal those of other residues.
+*Intra*-molecular charge constraints, or charge constraints that only apply to one molecule, are useful for enforcing equivalent charges between symmetric groups. For example, these are used to symmetrize the charges of hydrogens around a methyl or methylene group. They also allow for computing charges for a molecule that is compatible with an existing environment. One use case is calculating the charges of a non-canonical amino acid (ncAA) in a protein where the charges of canonical residues are already known; here, it is general practice to constrain the charges on the backbone of the ncAA to equal the known charges of the backbone in other residues.
 
-Inter-molecular charge constraints are particularly useful for computing the charges of multiple novel components of a molecule. For example, a polymer can be comprised of multiple monomer species. In order to calculate suitable charges for each monomer, their surrounding environment (the adjacent monomers) must be taken into account. Without charge constraints, the charges calculated for the same monomer in two different environments might differ. Inter-molecular charge constraints can be applied to constrain the charges of a particular monomer across all environments to be equivalent, ensuring that the derived charges can be used transferrably.
+*Inter*-molecular charge constraints, or charge constraints that apply to multiple molecules, are particularly useful for computing the charges of multiple components of a single macromolecule. For example, a co-polymer can be comprised of multiple monomer species, A and B. In order to calculate suitable charges for each monomer, their surrounding environment (the adjacent monomers) must be taken into account. However, the same monomer may be multiple different local environments: AAA, AAB, or BAB. Without charge constraints, the charges calculated for monomer A may differ between different environments, which lowers the transferability of the charge set for creating new polymers. Inter-molecular charge constraints can be applied to constrain the charges of a particular monomer across all environments to be equivalent, ensuring that the derived charges can be used transferrably. 
 
[…]
