+++ docs/source/examples/README.rst
@@ -7,9 +7,8 @@ Examples
 Here is a collection of example Jupyter notebooks.
 Each notebook can be downloaded to run on your own device.
 
-
 .. toctree::
    :maxdepth: 1
+   :glob:
 
-   resp_single
-   resp_multi
+   resp*
