+++ paper/paper.md
@@ -37,7 +37,9 @@ A number of methods have been developed to generate molecular partial charges, a
 The electrostatic potential $V$ is the potential generated by the nuclei and electrons in a molecule. At any point in space $r$, it is:
 $$V(\vec{r}) = \sum_{n=1}^{nuclei} \frac{Z_n}{|\vec{r} - \vec{R}_n|} - \int \frac{\rho(\vec{r’})}{|\vec{r} - \vec{r’}|} d\vec{r}$$
 
-The “ESP” method published by Singh and Kollman in 1984 [@singh1984] evaluates the electrostatic potential on a grid of points around the molecule. This approach is still used by the Automated Topology Builder [@malde2011] to derive charges for GROMOS. However, it is difficult to distinguish the contribution of buried atoms within the molecule to the electrostatic potential at the surface.  As a result, the charges assigned to the buried atoms can vary substantially with minor changes in geometry. 
+The “ESP” method published by Singh and Kollman in 1984 [@singh1984] evaluates the electrostatic potential on a grid of points around the molecule. This method canonically uses the HF/6-31G* level of theory in the gas phase. The approach is still used by the Automated Topology Builder [@malde2011] to derive charges for GROMOS, although the ATB method uses B3LYP/6-31G* in implicit solvent.
+
+One flaw of the ESP method is that it is difficult to distinguish the contribution of buried atoms within the molecule to the electrostatic potential at the surface.  As a result, the charges assigned to the buried atoms can vary substantially with minor changes in geometry. 
 
 ## RESP
 
@@ -45,11 +47,13 @@ The restrained electrostatic potential (RESP) [@bayly1993; @cornell1993; @ciepla
 
 $$\chi_{penalty} = a\sum_{n=1}^{nuclei} ((q_{n}^{2} + b^2)^{1/2} – b)$$
 
-$a$ defines the asymptotic limits of the penalty, while $b$ controls its width, or its degree of curvature at the minimum. $b$ is typically set to 0.1 e. In a two-stage fit, $a$ is usually tightened (increased) in the second stage. The standard values are 0.0005 au for the first stage, and 0.001 for the second stage. Hydrogens are often excluded in the hyperbolic penalty. Two-stage RESP is the typical charge model employed by the AMBER and GLYCAM force fields. 
+$a$ defines the asymptotic limits of the penalty, while $b$ controls its width, or its degree of curvature at the minimum. $b$ is typically set to 0.1 e. In a two-stage fit, $a$ is usually tightened (increased) in the second stage. The standard values are 0.0005 au for the first stage, and 0.001 for the second stage. Hydrogens are often excluded in the hyperbolic penalty. Two-stage RESP is the typical charge model employed by the AMBER and GLYCAM force fields.
+
+As with the ESP method, RESP typically computes the electrostatic potential at HF/6-31G* in the gas phase.
 
 ## RESP2
 
-The HF/6-31G* level of theory commonly used for RESP charge derivation overestimates the gas-phase polarity of molecules [@carlson1993]. The recently published RESP2 [@schauperl2020] addresses this issue by using a higher level of QM theory, as well as combining charges fitted to both gas- and aqueous-phase data. While optimising charge parameters alone without modifying Lennard-Jones parameters may not increase the accuracy of simulations [@mobley2007], both parameters can be co-optimised for greater accuracy. 
[…]
