+++ docs/source/examples/README.rst
@@ -11,4 +11,4 @@ Each notebook can be downloaded to run on your own device.
    :maxdepth: 1
    :glob:
 
-   example*
+   *example*
