### 2022-02-18 #4100 @ptmerz
Two comments regarding version:

* @lilyminium and @pdebuyl: The version given in the first comment of this review is v0.2.4. The version we are reviewing is 0.3.1. I think this was fixed in the pre-review, but this change doesn't seem to have been carried over.
* @lilyminium: The version string of the readthedocs documentation *document title* (the bit that appears in the tab) is 0.1-unstable.
