### 2022-02-20 #4100 @lilyminium
Thanks for pointing that out, @ptmerz! The [latest](https://psiresp.readthedocs.io/en/latest/) docs now properly specify the version as `psiresp v0.3.1+8.g4147f7c.dirty documentation` and `stable` will update on the next release :-)
