### 2022-02-22 lilyminium/psiresp#72 @ptmerz
Clarify the non-mandatory dependencies
I think the non-mandatory dependencies (RDKit, Psi4, and QCFractal) are a bit confusing. The documentation states

> The above dependencies can be difficult to install and unnecessary for PsiRESP’s core functionality.

However, none of the examples provided (not even the one titled "Minimal Example") can run without at least RDKit (which might be due to the use of SMILES codes?). Also, when trying to use functionality which would require RDKit, the error message reads

> ImportError: rdkit is a core dependency of PsiRESP. Please install it with `conda install -c conda-forge rdkit`

This error message ("core dependency") seems to be contradicting the earlier statement in the documentation.

My suggestion would be to either
* outline in the documentation what can be done without these dependencies, and maybe include an example, or,
* make them actual core dependencies, and add a note when installing through PyPI that users should install it themselves if they can't use conda.

Ref: JOSS review at https://github.com/openjournals/joss-reviews/issues/4100
