### 2022-02-22 lilyminium/psiresp#73 @ptmerz
Potential issue displaying job progress
This is definitely minor, but I thought I'd mention it in case it's an easy fix. When running the examples in the documentation (both for the single and the multi molecular case), when I would run the actual job,

```job.run(client=client)```

the output

```generate-conformers: 100%|█████████████████████████████| 1/1 [00:00<00:00, 276.27it/s]```

would appear immediately, followed by a long-ish break, after which the second output

```compute-esp: 100%|█████████████████████████████| 1/1 [00:00<00:00, 16.75it/s]```

would appear, immediately at 100%. I was wondering whether
* this is due to the simple example, and the bar would appear gradually for more complex examples, or
* if there is a small bug which keeps the bar from being filled gradually over the calculation, or
* if there is another long step in between the two which does not display any progress information.

I observed this running the examples both in a Jupyter notebook and directly in a Python terminal.

Ref: JOSS review at https://github.com/openjournals/joss-reviews/issues/4100
