### 2022-02-22 #4100 @ptmerz
I thoroughly enjoyed reviewing the package and the paper. Congratulations on a great job @lilyminium!

I am convinced the package will be useful to the community. It addresses a very common problem in molecular simulations. It is feature-rich, very well documented, intuitive to use, and follows Python research software best practices.

The paper is well written and clearly outlines both the need for the software and it's functionality.

All my raised issues have been minor, and most have already been taken care of. I have ticked all review boxes above, since the last remaining issues are extremely minor and I definitely would not see them as blocking issues. Note that I have already ticked the "Community Guidelines" tickbox since the guidelines are about to be introduced in https://github.com/lilyminium/psiresp/pull/70.

I think the software and the paper are a great fit for JOSS and I can confidently recommend it for acceptance.
