### 2022-02-27 e9c01c7
make example more minimal
