### 2022-02-27 7a9ef3a
Add tutorial, update paper, link to examples (#70)
