+++ release_history.md
@@ -4,6 +4,18 @@ Releases follow the `major.minor.micro` format of [PEP440](https://www.python.or
 
 ## Current development
 
+- Add progress bars (#75)
+- Add tutorial for using psiresp-base (#74)
+- Add additional tutorial for RESP without QCFractal server (#70)
+
+## 0.3.1
+-  Add explicit check for singular matrices (PR #65)
+
+## 0.3.0
+
+- Split out dependencies (PR #60)
+- Rename core attributes (PR #59)
+- Added more tests (PR #50)
 - Show inherited documentation (PR #48)
 - Add contributions, dependencies, references (PR #44)
 - Fix documentation rendering (PR #37)
