### 2022-02-27 lilyminium/psiresp#73 @lilyminium
Thanks @ptmerz! You're right, there are some long steps where QCFractal is waiting for Psi4 to complete, which were not given progress bars. Thanks to the great tqdm library it is fairly easy to fix this, and I've added progress bars in #75. As you've said that this is not a blocking problem, I will merge as-is so as not to take up more of your time, but please do let me know if this is not the best solution!
