### 2022-02-27 lilyminium/psiresp#72 @lilyminium
Thanks for noticing that inconsistency and raising this, @ptmerz. RDKit and Psi4 did use to be core dependencies, but it is possible to use PsiRESP without them if you have geometries, grids, and ESPs from another source -- I added a new tutorial in #74 to clarify that.
