### 2022-02-27 5f7a8f4
Add minimal tutorial (#74)
