### 2022-02-27 80bbe85
Add minimal tutorial (#74)
