### 2022-02-28 f51cce8
Remove redundant contribution section
