### 2022-03-05 85ffa8c
add __init__ signature to Model
