+++ psiresp/base.py
@@ -1,5 +1,4 @@
 import inspect
-import functools
 import hashlib
 from typing import Any, Optional, Union, no_type_check
 
@@ -23,7 +22,6 @@ def _to_immutable(obj):
     return obj
 
 
-
 class Model(BaseModel):
     """Base class that all classes should subclass.
     """
@@ -38,10 +36,10 @@ class Model(BaseModel):
     def _clsname(self):
         return type(self).__name__
 
-    def __init__(__pydantic_self__, **data: Any) -> None:
-        __pydantic_self__.__pre_init__(**data)
+    def __init__(__pydantic_self__, **data: Any) -> None:  # lgtm[py/not-named-self]
+        __pydantic_self__.__pre_init__(**data)  # lgtm[py/init-calls-subclass]
         super().__init__(**data)
-        __pydantic_self__.__post_init__(**data)
+        __pydantic_self__.__post_init__(**data)  # lgtm[py/init-calls-subclass]
 
     def __pre_init__(self, **kwargs):
         pass
