### 2022-03-05 6f63fc5
Add __init__ signature to Model (#79)
Fixes #78
