+++ psiresp/tests/data/manual_jobs/psiresp_working_directory/optimization/run_optimization.sh
@@ -1,3 +1,3 @@
-psi4 --qcschema psiresp_working_directory/optimization/CH6N_795ee2a77c6a4347fd76ed1ab0f8c486d24a2238_c9ce731306cb83b137c5cfd5f69a120483b61005.msgpack
-psi4 --qcschema psiresp_working_directory/optimization/C7H14N2O2_9ee96ceb2aec1b0d4b5c53ad3ae9e61d546f6717_c9ce731306cb83b137c5cfd5f69a120483b61005.msgpack
-psi4 --qcschema psiresp_working_directory/optimization/C7H14N2O2_abb25794aba793b7bf575666eaefae61736f254e_c9ce731306cb83b137c5cfd5f69a120483b61005.msgpack
+psi4 --qcschema psiresp_working_directory/optimization/CH6N_795ee2a77c6a4347fd76ed1ab0f8c486d24a2238_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
+psi4 --qcschema psiresp_working_directory/optimization/C7H14N2O2_9ee96ceb2aec1b0d4b5c53ad3ae9e61d546f6717_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
+psi4 --qcschema psiresp_working_directory/optimization/C7H14N2O2_abb25794aba793b7bf575666eaefae61736f254e_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
