+++ psiresp/tests/data/manual_jobs/psiresp_working_directory/single_point/run_single_point.sh
@@ -1,10 +1,10 @@
-psi4 --qcschema psiresp_working_directory/single_point/CH6N_5522973281a29d00873575078945db705e9e0167_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/CH6N_ecb8b34e0b75ae8e0a73bc53bdd6a4d5c1a5b5c2_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_607de52ef03791820ad946978d48703837b9338c_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_4d7331ba8d65f69230c53e612f3bc259271675a0_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_da2a0eaa440175fbcb099bd6b74de7ac980c9b50_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_0c7e913c4a8462f7181fc25edf913be91be8d7c4_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_5536bc6a52953c07b31bf82f85c8e90f2142cccf_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_3ec849bebba8d4b8054b40a889b681d861afc28b_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_5899f84c638c2be4f9e3ba4cd9beedff56c6cc3c_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
-psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_daaa75b70b8b2b50a22d1441439a1affcb9be48d_e746222796fc2c4c5a1f896fa1cc1cefffe7044c.msgpack
+psi4 --qcschema psiresp_working_directory/single_point/CH6N_5522973281a29d00873575078945db705e9e0167_4d9bb6d245f75e8aea535e7b9676614d130d864b.msgpack
+psi4 --qcschema psiresp_working_directory/single_point/CH6N_ecb8b34e0b75ae8e0a73bc53bdd6a4d5c1a5b5c2_4d9bb6d245f75e8aea535e7b9676614d130d864b.msgpack
+psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_607de52ef03791820ad946978d48703837b9338c_4d9bb6d245f75e8aea535e7b9676614d130d864b.msgpack
+psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_4d7331ba8d65f69230c53e612f3bc259271675a0_4d9bb6d245f75e8aea535e7b9676614d130d864b.msgpack
+psi4 --qcschema psiresp_working_directory/single_point/C7H14N2O2_da2a0eaa440175fbcb099bd6b74de7ac980c9b50_4d9bb6d245f75e8aea535e7b9676614d130d864b.msgpack
[…]
