### 2022-03-08 #4100 @pdebuyl
Hi @lilyminium I have some issues with the pip install, I'll get back to you when it is solved. It is not an issue with PsiRESP per se indeed, but given the choice I always try pip first and the postgresql routines in qcfractal are not portable.
