### 2022-03-10 #4100 @lilyminium
@pdebuyl rats, that's annoying. I think QCFractal does have some notes somewhere that installation by pip is not optimal; perhaps it's related. Which architecture are you on? Do you find that it works with conda? I just ask because the `conda install psiresp` package should functionally do a `pip install psiresp`, followed by conda installing the remaining dependencies.
