### 2022-03-10 #4100 @pdebuyl
I use debian. Googling postgresql initdb resulted in interesting finds about debian (and Ubuntu I believe) being different w.r.t. postgresql setup. I'll try once more and if it doesn't work go the conda route.
