### 2022-03-11 01c3705
add minimization
