### 2022-03-20 6e8069f
add minimization
