### 2022-03-20 b9fc6a7
Add conformer minimization (#83)
