+++ psiresp/rdutils.py
@@ -1,7 +1,8 @@
 
 import itertools
 import logging
-from typing import TYPE_CHECKING, Set, Tuple
+from typing import TYPE_CHECKING, Set, Tuple, Optional
+import warnings
 
 from typing_extensions import Literal
 import numpy as np
@@ -27,6 +28,7 @@ BONDTYPES = {
 
 logger = logging.getLogger(__name__)
 
+OPT_FF_NAMES = Literal["uff", "mmff94"]
 OFF_SMILES_ATTRIBUTE = "canonical_isomeric_explicit_hydrogen_mapped_smiles"
 
 
@@ -246,7 +248,9 @@ def add_conformer_from_coordinates(rdmol: "rdkit.Chem.Mol",
 def generate_conformers(rdmol: "rdkit.Chem.Mol",
                         n_conformers: int = 0,
                         rms_tolerance: float = 1.5,
-                        random_seed=-1):
+                        random_seed=1,
+                        minimize: Optional[OPT_FF_NAMES] = None,
+                        ):
     """Generate conformers for an RDKit molecule.
 
     This does not clear existing conformers.
@@ -267,9 +271,12 @@ def generate_conformers(rdmol: "rdkit.Chem.Mol",
                                clearConfs=False,
                                randomSeed=random_seed,
                                ignoreSmoothingFailures=True)
+    if minimize:
+        minimize_conformer_geometries(rdmol, forcefield=minimize)
 
 
 def minimize_conformer_geometries(rdmol: "rdkit.Chem.Mol",
+                                  forcefield: OPT_FF_NAMES = "mmff94",
                                   minimize_max_iter: int = 2000):
     """Minimize conformer geometries of an RDKit molecule
 
@@ -278,10 +285,20 @@ def minimize_conformer_geometries(rdmol: "rdkit.Chem.Mol",
     minimize_max_iter: int (optional)
         Maximum number of iterations for minimization
     """
+    # TODO: make this more customizable
 
-    # TODO: is UFF good?
-    AllChem.UFFOptimizeMoleculeConfs(rdmol, numThreads=0,
-                                     maxIters=minimize_max_iter)
+    forcefield = forcefield.lower()
+    if forcefield == "uff":
+        AllChem.UFFOptimizeMoleculeConfs(rdmol, numThreads=0,
+                                         maxIters=minimize_max_iter)
+    elif forcefield == "mmff94":
+        AllChem.MMFFOptimizeMoleculeConfs(rdmol, numThreads=0,
+                                          maxIters=minimize_max_iter)
+    else:
+        raise ValueError(
+            f"Force field {forcefield} not supported. "
+            f"Please use one of {OPT_FF_NAMES}"
+        )
 
 
 def compute_mmff_charges(rdmol: "rdkit.Chem.Mol",
@@ -366,8 +383,13 @@ def select_elf_conformer_ids(rdmol: "rdkit.Chem.Mol",
     if n_conformers == 0:
         return
 
-    energies = compute_electrostatic_energy(rdmol, forcefield="MMFF94")
     all_conformer_ids = [c.GetId() for c in rdmol.GetConformers()]
+
+    try:
+        energies = compute_electrostatic_energy(rdmol, forcefield="MMFF94")
+    except ValueError as e:
+        warnings.warn(str(e))
+        return all_conformer_ids[:limit]
 
     sorting = np.argsort(energies)
     conversion = qcel.constants.conversion_factor(
@@ -408,6 +430,7 @@ def generate_diverse_conformer_coordinates(molecule,
                                            energy_window: float = 30,
                                            n_max_conformers: int = 10,
                                            rms_tolerance: float = 0.05,
[…]
