+++ psiresp/tests/datafiles.py
@@ -65,3 +65,5 @@ METHYLAMMONIUM_O1_GRID = datapath("esps/27c00e84078b824244612107da714196f49424cd
 METHYLAMMONIUM_O1_ESP = datapath("esps/27c00e84078b824244612107da714196f49424cd_esp.dat")
 METHYLAMMONIUM_O2_GRID = datapath("esps/8776ca97115f78967cccc3dbc89c5530891000db_grid.dat")
 METHYLAMMONIUM_O2_ESP = datapath("esps/8776ca97115f78967cccc3dbc89c5530891000db_esp.dat")
+
+TRIFLUOROETHANOL_JOB = datapath("jobs/trifluoroethanol.json")
