### 2022-03-21 2861a91
add symmetry test
