### 2022-03-22 bd0a1de
move restraints and make options available to user
