+++ psiresp/charge.py
@@ -59,8 +59,8 @@ class BaseChargeConstraint(base.Model):
                                  molecule_increments: Dict[int, int] = {}):
         return self.to_sparse_row_constraint(n_dim, molecule_increments).transpose()
 
-    def get_atom_indices(self, molecule_increments: Dict[int, int] = {}):
-        indices = [atom.index + molecule_increments.get(hash(atom.molecule), 0)
+    def get_atom_indices(self, molecule_increments: Dict[int, List[int]] = {}):
+        indices = [atom.index + molecule_increments.get(hash(atom.molecule), [0])[0]
                    for atom in self.atoms]
         return np.array(sorted(indices), dtype=int)
 
@@ -92,7 +92,7 @@ class ChargeSumConstraint(BaseChargeConstraint):
     #     )
 
     def to_row_constraint(self, n_dim: int,
-                          molecule_increments: Dict[Molecule, int] = {}):
+                          molecule_increments: Dict[int, List[int]] = {}):
         row = np.zeros((1, n_dim))
         indices = self.get_atom_indices(molecule_increments)
         row[0, indices] = 1
@@ -109,23 +109,38 @@ class ChargeEquivalenceConstraint(BaseChargeConstraint):
     """
 
     def to_row_constraint(self, n_dim: int,
-                          molecule_increments: Dict[Molecule, int] = {}):
-        n_items = len(self) - 1
-        rows = np.zeros((n_items, n_dim))
+                          molecule_increments: Dict[int, List[int]] = {}):
         indices = self.get_atom_indices(molecule_increments)
-        for i, (j, k) in enumerate(zip(indices[:-1], indices[1:])):
-            rows[i][j] = -1
-            rows[i][k] = 1
-        return rows
+        return self._convert_indices_to_constraint_rows(n_dim, indices)
 
     @property
     def charge(self):
         return None
 
+    @staticmethod
+    def _convert_indices_to_constraint_rows(n_dim, indices):
+        n_items = len(indices) - 1
+        rows = np.zeros((n_items, n_dim))
+        for i, (j, k) in enumerate(zip(indices[:-1], indices[1:])):
+            rows[i][j] = -1
+            rows[i][k] = 1
+        return rows
+
 
 class BaseChargeConstraintOptions(base.Model):
     charge_sum_constraints: List[ChargeSumConstraint] = []
     charge_equivalence_constraints: List[ChargeEquivalenceConstraint] = []
+    split_conformers: bool = Field(
+        default=False,
+        description="Treat conformers separately, instead of combining the restraint matrices"
+    )
+    constrain_methyl_hydrogens_between_conformers: bool = Field(
+        default=False,
+        description=(
+            "Whether to constrain methyl/ene hydrogens as equivalent between conformers. "
+            "This has no effect if `split_conformers=False`."
+        )
+    )
 
     @property
     def n_constraints(self):
@@ -157,8 +172,17 @@ class BaseChargeConstraintOptions(base.Model):
             for atom in chrequiv.atoms:
                 equivalences[atom] |= chrequiv.atoms
 
-        self.charge_equivalence_constraints = [ChargeEquivalenceConstraint(atoms=x)
-                                               for x in equivalences.values()]
+        unique_sets = []
+        for group in equivalences.values():
+            for existing in unique_sets:
+                if group & existing:
+                    existing |= group
+                    break
+            else:
+                unique_sets.append(group)
+
+        self.charge_equivalence_constraints = [ChargeEquivalenceConstraint(atoms=sorted(x))
+                                               for x in unique_sets]
 
     def _get_single_atom_charge_constraints(self) -> Dict[Atom, float]:
         """Get ChargeConstraints with only one atom as a dict"""
@@ -222,7 +246,11 @@ class BaseChargeConstraintOptions(base.Model):
         """
[…]
