### 2022-03-22 aa53255
add more tests and docs
