### 2022-03-22 2b20932
fix split conformer equivalences
