+++ psiresp/tests/test_job.py
@@ -106,8 +106,8 @@ class TestMultiRespWithoutClient:
         (True, 0.0005, AMM_NME_OPT_RESPA1_CHARGES),
     ], indirect=['red_charges'])
     def test_given_esps_split_conf(self, nme2ala2, methylammonium,
-                        methylammonium_nme2ala2_charge_constraints,
-                        stage_2, restraint_height, red_charges, job_esps, job_grids):
+                                   methylammonium_nme2ala2_charge_constraints,
+                                   stage_2, restraint_height, red_charges, job_esps, job_grids):
 
         resp_options = RespOptions(stage_2=stage_2, restraint_height_stage_1=restraint_height)
         job = Job(molecules=[methylammonium, nme2ala2],
@@ -131,7 +131,6 @@ class TestMultiRespWithoutClient:
         assert_allclose(charges[18], charges[22])
         for calculated, reference in zip(job.charges[::-1], red_charges[::-1]):
             assert_allclose(calculated, reference, atol=1e-3)
-
 
     @requires_qcfractal
     def test_run_manual(self, nme2ala2_empty, methylammonium_empty, tmpdir):
