### 2022-03-22 86c1ebf
Make options for splitting conformers and equivalencing Hs available to the user (#85)
