+++ psiresp/tests/test_charge.py
@@ -15,6 +15,8 @@ from psiresp.constraint import SparseGlobalConstraintMatrix
 from psiresp.tests.datafiles import (
     DMSO_STAGE_2_A, DMSO_STAGE_2_B,
     DMSO_JOB_WITH_ORIENTATION_ENERGIES,
+    TRIFLUOROETHANOL_JOB,
+    AMM_NME_RESP_JOB
 )
 
 
@@ -26,7 +28,7 @@ def test_charge_sum_constraint(dmso):
     assert_equal(constraint.molecules, np.array([dmso, dmso]))
     assert constraint.molecule_set == {dmso}
 
-    molinc = {hash(dmso): 4}
+    molinc = {hash(dmso): [4]}
     indices = constraint.get_atom_indices(molecule_increments=molinc)
     assert_allclose(indices, [5, 6])
     row = constraint.to_row_constraint(10, molecule_increments=molinc)
@@ -41,7 +43,7 @@ def test_charge_equivalence_constraint(dmso):
     assert_equal(constraint.molecules, np.array([dmso, dmso, dmso]))
     assert constraint.molecule_set == {dmso}
 
-    molinc = {hash(dmso): 4}
+    molinc = {hash(dmso): [4]}
     indices = constraint.get_atom_indices(molecule_increments=molinc)
     assert_allclose(indices, [5, 6, 7])
     row = constraint.to_row_constraint(10, molecule_increments=molinc)
@@ -103,6 +105,7 @@ class TestMoleculeChargeConstraints:
         assert len(constraints.charge_equivalence_constraints) == 2
 
         surface_constraints = job.construct_surface_constraint_matrix()
+        assert surface_constraints.matrix.shape == (12, 11)
         matrix = SparseGlobalConstraintMatrix.from_constraints(surface_constraints,
                                                                constraints)
 
@@ -111,3 +114,24 @@ class TestMoleculeChargeConstraints:
 
         assert_allclose(matrix.coefficient_matrix.toarray(), ref_a)
         assert_allclose(matrix.constant_vector, ref_b)
+
+
+class TestSurfaceConstraintMatrix:
+    @pytest.mark.parametrize("jobfile, n_mol, n_confs, nosplit, split", [
+        (DMSO_JOB_WITH_ORIENTATION_ENERGIES, 1, [1], (12, 11), (12, 11)),
+        (TRIFLUOROETHANOL_JOB, 1, [10], (14, 13), (131, 130)),
+        (AMM_NME_RESP_JOB, 2, [1, 2], (36, 35), (62, 61)),
+    ])
+    def test_conformer_splitting_one_conf(self, jobfile, n_mol, n_confs, nosplit, split):
+        job = Job.parse_file(jobfile)
+        charge_options = job.charge_constraints
+        assert len(job.molecules) == n_mol
+        assert [mol.n_conformers for mol in job.molecules] == n_confs
+
+        charge_options.split_conformers = False
+        surface_constraints = job.construct_surface_constraint_matrix()
+        assert surface_constraints.matrix.shape == nosplit
+
+        charge_options.split_conformers = True
+        surface_constraints = job.construct_surface_constraint_matrix()
+        assert surface_constraints.matrix.shape == split
