+++ psiresp/job.py
@@ -232,20 +232,40 @@ class Job(base.Model):
         Construct the constraint matrix for each atom,
         as generated by the ESP at each grid point
         """
-        matrices = [
-            ESPSurfaceConstraintMatrix.from_orientations(
-                orientations=[o for conf in mol.conformers for o in conf.orientations],
-                temperature=self.temperature,
-            )
-            for mol in self.molecules
-        ]
+        if not self.charge_constraints.split_conformers:
+            matrices = [
+                ESPSurfaceConstraintMatrix.from_orientations(
+                    orientations=[o for conf in mol.conformers for o in conf.orientations],
+                    temperature=self.temperature,
+                )
+                for mol in self.molecules
+            ]
+            charges = [mol.charge for mol in self.molecules]
+        else:
+            matrices = [
+                ESPSurfaceConstraintMatrix.from_orientations(
+                    orientations=[o for o in conf.orientations],
+                    temperature=self.temperature,
+                )
+                for mol in self.molecules
+                for conf in mol.conformers
+            ]
+            charges = [
+                chg
+                for mol in self.molecules
+                for chg in [mol.charge] * mol.n_conformers
+            ]
+
         a_mol = scipy.linalg.block_diag(*[mat.coefficient_matrix for mat in matrices])
         a_row = scipy.linalg.block_diag(*[np.ones(mat.constant_vector.shape[0]) for mat in matrices])
         a_zeros = np.zeros((a_row.shape[0], a_row.shape[0]))
         a_block = np.bmat([[a_mol, a_row.T], [a_row, a_zeros]])
 
-        b_block = np.concatenate([mat.constant_vector for mat in matrices]
-                                 + [[mol.charge for mol in self.molecules]])
+        b_vecs = [mat.constant_vector for mat in matrices] + [charges]
+        if len(b_vecs):
+            b_block = np.concatenate(b_vecs)
+        else:
+            b_block = np.zeros((0, 0))
         return ESPSurfaceConstraintMatrix.from_coefficient_matrix(a_block, b_block)
 
     def generate_molecule_charge_constraints(self) -> MoleculeChargeConstraints:
@@ -266,7 +286,10 @@ class Job(base.Model):
 
         if self.resp_options.stage_2:
             stage_2_constraints = stage_1_constraints.copy(deep=True)
-            stage_1_constraints.charge_equivalence_constraints = []
+            stage_2_constraints.constrain_methyl_hydrogens_between_conformers = True
+            stage_1_constraints.prepare_stage_1_constraints()
+        else:
+            stage_1_constraints.constrain_methyl_hydrogens_between_conformers = True
 
         self.stage_1_charges = RespCharges(charge_constraints=stage_1_constraints,
                                            surface_constraints=surface_constraints,
@@ -275,6 +298,7 @@ class Job(base.Model):
         self.stage_1_charges.solve()
 
         if self.resp_options.stage_2:
+            stage_2_constraints.prepare_stage_2_constraints()
             stage_2_constraints.add_constraints_from_charges(self.stage_1_charges._charges)
             self.stage_2_charges = RespCharges(charge_constraints=stage_2_constraints,
                                                surface_constraints=surface_constraints,
