### 2022-03-25 fb92d1c
make basis set more free
