### 2022-03-25 #4100 @pdebuyl
I ended up using the conda install. The install page makes it explicit that it is easier but I had a package resolution conflict as well by doing `conda install -c conda-forge psiresp`. I used the file `test_env.yml` in the end and that worked.

The install instructions also state that it is possible to install with `python setup.py install` (after the dependencies have been installed). That made an "egg" installed that did not bring the data. Could you change that part to recommend `pip install .` instead? That worked ok (with conda-installed environment).
