### 2022-03-25 #4100 @pdebuyl
References:
- Would you have a reference (link to documentation or homepage of the code if there is no paper) for amber and gaussian?
- Can you clean the entry for rdkit?
- for schauperl 2020 there is an article number 44
