### 2022-03-27 6523b97
change python setup install to pip install
