### 2022-03-27 1e34590
remove files and clean up amber citation
