+++ paper/paper.bib
@@ -284,7 +284,7 @@
 }
 
 @software{case,
-  title     = {Amber 2020},
+  title     = {Amber 2021},
   author    = {D.A. Case
                and Aktulga, H.M.
                and Belfon, K.
@@ -349,7 +349,7 @@
                and Zhao, S.
                and Kollman, P.A.
                },
-  year      = {2020},
+  year      = {2021},
   publisher = {University of California, San Francisco}
 }
 
