+++ .github/workflows/gh-ci.yaml
@@ -78,7 +78,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
           auto-update-conda: true
           channel-priority: flexible
-          channels: defaults, conda-forge, psi4
+          channels: conda-forge, psi4
           add-pip-as-python-dependency: true
           mamba-version: "*"
           architecture: x64
