+++ .github/workflows/gh-ci.yaml
@@ -76,9 +76,9 @@ jobs:
         uses: conda-incubator/setup-miniconda@v2
         with:
           python-version: ${{ matrix.python-version }}
-          auto-update-conda: true
+          auto-update-conda: false
           channel-priority: flexible
-          channels: conda-forge, psi4
+          channels: defaults, conda-forge, psi4
           add-pip-as-python-dependency: true
           mamba-version: "*"
           architecture: x64
