+++ .github/workflows/gh-ci.yaml
@@ -76,7 +76,7 @@ jobs:
         uses: conda-incubator/setup-miniconda@v2
         with:
           python-version: ${{ matrix.python-version }}
-          auto-update-conda: false
+          auto-update-conda: true
           channel-priority: flexible
           channels: defaults, conda-forge, psi4
           add-pip-as-python-dependency: true
