+++ devtools/conda-envs/no_qcfractal_test_env.yaml
@@ -18,7 +18,7 @@ dependencies:
 
   # qcstack
   - postgresql >= 13.0
-  - qcelemental == 0.24.0
+  - qcelemental
   - requests
 
   # deps
