+++ paper/paper.bib
@@ -16,8 +16,7 @@
   year       = {2011},
   note       = {00670},
   keywords   = {Must-cite, MDAnalysis},
-  pages      = {2319--2327},
-  file       = {Michaud-Agrawal_2011_MDAnalysis_-_A_toolkit_for_the_analysis_of_molecular_dynamics_simulations.pdf:/Users/lily/OneDrive - Australian National University/Papers/Michaud-Agrawal_2011_MDAnalysis_-_A_toolkit_for_the_analysis_of_molecular_dynamics_simulations2.pdf:application/pdf;Michaud-Agrawal_2011_MDAnalysis_-_A_toolkit_for_the_analysis_of_molecular_dynamics_simulations.pdf:/Users/lily/OneDrive - Australian National University/Papers/Michaud-Agrawal_2011_MDAnalysis_-_A_toolkit_for_the_analysis_of_molecular_dynamics_simulations.pdf:application/pdf;Michaud‐Agrawal_2011_MDAnalysis.pdf:/Users/lily/OneDrive - Australian National University/Papers/Tools and data/Michaud‐Agrawal_2011_MDAnalysis.pdf:application/pdf;Snapshot:/Users/lily/Zotero/storage/PKBTSUWN/jcc.html:text/html}
+  pages      = {2319--2327}
 }
 
 @article{gowers2016,
@@ -31,8 +30,7 @@
   year       = {2016},
   note       = {00152},
   keywords   = {Must-cite, MDAnalysis},
-  pages      = {98--105},
-  file       = {Snapshot:/Users/lily/Zotero/storage/586SRBNY/oliver_beckstein.html:text/html;Gowers_2016_MDAnalysis_-_A_Python_Package_for_the_Rapid_Analysis_of_Molecular_Dynamics_Simulations.pdf:/Users/lily/OneDrive - Australian National University/Papers/Gowers_2016_MDAnalysis_-_A_Python_Package_for_the_Rapid_Analysis_of_Molecular_Dynamics_Simulations2.pdf:application/pdf;Snapshot:/Users/lily/Zotero/storage/72E5XCCY/oliver_beckstein.html:text/html;Snapshot:/Users/lily/Zotero/storage/MCIY2VSG/oliver_beckstein.html:text/html;Gowers_2016_MDAnalysis.pdf:/Users/lily/OneDrive - Australian National University/Papers/Tools and data/Gowers_2016_MDAnalysis.pdf:application/pdf}
+  pages      = {98--105}
 }
 
 @article{malde2011,
@@ -52,8 +50,7 @@
   year       = {2011},
   note       = {00000},
   keywords   = {Must-cite, ATB},
-  pages      = {4026--4037},
-  file       = {Malde_2011_An_Automated_Force_Field_Topology_Builder_(ATB)_and_Repository_-_Version_1.0.pdf:/Users/lily/OneDrive - Australian National University/Papers/Malde_2011_An_Automated_Force_Field_Topology_Builder_(ATB)_and_Repository_-_Version_1.0.pdf:application/pdf}
+  pages      = {4026--4037}
 }
 
 
@@ -73,8 +70,7 @@
   year       = {1995},
   note       = {00876},
   keywords   = {RESP},
-  pages      = {1357--1377},
-  file       = {Cieplak_1995_Application_of_the_multimolecule_and_multiconformational_RESP_methodology_to_biopolymers_-_Charge.pdf:/Users/lily/OneDrive - Australian National University/Papers/Cieplak_1995_Application_of_the_multimolecule_and_multiconformational_RESP_methodology_to_biopolymers_-_Charge.pdf:application/pdf}
+  pages      = {1357--1377}
 }
 
 @article{cornell1993,
[…]
