+++ .github/workflows/gh-ci.yaml
@@ -80,7 +80,7 @@ jobs:
           channel-priority: flexible
           channels: defaults, conda-forge, psi4
           add-pip-as-python-dependency: true
-          mamba-version: "*"
+          # mamba-version: "*"
           architecture: x64
           environment-file: devtools/conda-envs/${{ matrix.environment }}.yaml
           activate-environment: test
