### 2022-03-28 4e0fd88
Fix installation instructions (#86)
