### 2022-03-28 da7abae
Fix documentation dependencies (#89)
