+++ devtools/conda-envs/docs_env.yaml
@@ -11,13 +11,13 @@ dependencies:
 
   # qm
   - dftd3
-  - psi4 >= 1.3.0
+  - psi4 < 1.5
   - intel-openmp!=2019.5
   - geometric
 
   # qcstack
   - postgresql >= 13.0
-  - qcelemental==0.22.0
+  - qcelemental > 0.23
   - qcengine
   - qcfractal
   - qcfractal-core
@@ -35,6 +35,7 @@ dependencies:
   - jupyter
   - nglview
   - nbsphinx
+  - jinja2 < 3.1  # https://github.com/pallets/jinja/pull/1544
   - sphinx > 2.0
 
   # Pip-only installs
