### 2022-03-28 #4100 @lilyminium
@pdebuyl yikes, would you recall what the package conflict was? Was it anything to do with `qcfractal`? I'm trying to reproduce on Ubuntu but haven't found anything yet so far.

> The install instructions also state that it is possible to install with python setup.py install (after the dependencies have been installed). That made an "egg" installed that did not bring the data. Could you change that part to recommend pip install . instead? That worked ok (with conda-installed environment).

Thanks, done. Sorry, I was under the impression they were functionally the same thing.

> Please fix the invalid DOI

I think that's supposed to be `10.1002/(SICI)1096-987X(20000130)21:2<132::AID-JCC5>3.0.CO;2-P`, I'm not sure if the bot strips punctuation when it checks. I believe that is what I have in the `paper.bib` file and it is valid: [https://doi.org/10.1002/(SICI)1096-987X(20000130)21:2<132::AID-JCC5>3.0.CO;2-P](https://doi.org/10.1002/(SICI)1096-987X(20000130)21:2<132::AID-JCC5>3.0.CO;2-P)

Here it is in the master branch currently: https://github.com/lilyminium/psiresp/blob/86c1ebf228869ef835933ae5d9bcaff48f0d3d59/paper/paper.bib#L248

> Would you have a reference (link to documentation or homepage of the code if there is no paper) for amber and gaussian?

I think these are already in the paper and cited -- I did clean up the AMBER citation as it was not showing all authors, and change it to 2021. So far as I can tell these are the desired citations.

They are cited in the passages here -- I removed subsequent citations.

> distributed by the AmberTools package (Case et al., 2020).
> It supports quantum chemistry jobs using both the Gaussian (Frisch et al., 2016)

> Can you clean the entry for rdkit?

Do you mean cleaning up the author vs user names? I am happy to, but I'm pretty new to this and not sure if I should -- the entry for RDKit comes from this [zenodo entry](https://zenodo.org/record/4107869) and has a reasonably unambiguous Bibtex entry / requested citation already. Would that mess anything up for them?

> for schauperl 2020 there is an article number 44

Thanks -- I've added this as a `pages = {44}`. This seems to be the most common way to represent this in bibtex, where there's apparently not a canonical field. Please let me know if it should be something else.
