### 2022-03-28 #4100 @pdebuyl
Ok, I retried:

```
conda create -n tmpenv python=3
conda activate tmpenv
conda install -c conda-forge psiresp
conda install notebook
```

Installing psiresp triggered Python 3.8 to be selected.

Running example `01_examples_resp_single_server.ipynb` returns

``` py
ImportError                               Traceback (most recent call last)
Input In [5], in <cell line: 1>()
----> 1 from qcfractal import FractalSnowflakeHandler
      3 server = FractalSnowflakeHandler()
      4 print(server)

File /opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/qcfractal/__init__.py:5, in <module>
      1 """
      2 Main init function for qcfractal
      3 """
----> 5 from . import interface
      6 # Handle versioneer
      7 from .extras import get_information

File /opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/qcfractal/interface/__init__.py:5, in <module>
      1 """
      2 DQM Client base folder
      3 """
----> 5 from . import collections, data, models, util
      6 # Add imports here
      7 from .client import FractalClient

File /opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/qcfractal/interface/collections/__init__.py:6, in <module>
      1 """
      2 Python init for QCPortal collections
      3 """
      5 from .collection_utils import collection_factory, collections_name_map, list_known_collections, register_collection
----> 6 from .dataset import Dataset
      7 from .generic import Generic
      8 from .gridoptimization_dataset import GridOptimizationDataset

File /opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/qcfractal/interface/collections/dataset.py:12, in <module>
      8 import pandas as pd
     10 from qcelemental import constants
---> 12 from ..models import ComputeResponse, KeywordSet, Molecule, ObjectId, ProtoModel
     13 from ..statistics import wrap_statistics
     14 from ..visualization import bar_plot, violin_plot

File /opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/qcfractal/interface/models/__init__.py:5, in <module>
      1 """
      2 Either pull in QCEl models or local models
      3 """
----> 5 from . import rest_models
      6 from .common_models import (AutodocBaseSettings, KeywordSet, Molecule, ObjectId, OptimizationSpecification, ProtoModel,
      7                             QCSpecification)
      8 from .gridoptimization import GridOptimizationInput, GridOptimizationRecord

File /opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/qcfractal/interface/models/rest_models.py:6, in <module>
      1 """
      2 Models for the REST interface
      3 """
      4 from typing import Any, Dict, List, Optional, Tuple, Union
----> 6 from pydantic import Schema, constr, validator
      8 from qcelemental.util import get_base_docs
     10 from .common_models import KeywordSet, Molecule, ObjectId, ProtoModel

ImportError: cannot import name 'Schema' from 'pydantic' (/opt/miniconda3/envs/tmpenv/lib/python3.8/site-packages/pydantic/__init__.cpython-38-x86_64-linux-gnu.so)
```
