### 2022-03-28 #4100 @pdebuyl
The reference "jakalian2000" is not cited if I understand correctly, could you thus remove it from the bib file, as well as other non-cited entries?
