### 2022-03-28 #4100 @pdebuyl
rdkit appears as
```
Landrum, G., Tosco, P., Kelley, B., sriniker, gedeck, Ric, Vianello, R., NadineSchneider, Dalke,204
A., N, D., Cole, B., Kawashima, E., Turk, S., Swain, M., AlexanderSavelyev, Cosgrove,205
D., Vaucher, A., Wójcikowski, M., Probst, D., ... Jensen, J. H. (2020). Rdkit/rdkit:206
2020_09_1 (Q3 2020) release (Version Release_2020_09_1) [Computer software]. Zenodo.207
https://doi.org/10.5281/zenodo.4107869
```

the ` (Version Release_2020_09_1)` is not necessary, the "version" entry is duplicate with the title (only one could be kept). The `[Computer software]` part is apparently from our bibtex style, sorry about it.
