### 2022-03-28 #4100 @pdebuyl
Regarding Gaussian and Amber, those seem indeed to be the recommended citations, sorry as well. It is a bit weird from them but whatever :-)
