+++ paper/paper.bib
@@ -211,42 +211,6 @@
   pages     = {129--145}
 }
 
-@article{mulliken1955,
-  title   = {Electronic {Population} {Analysis} on {LCAO}–{MO} {Molecular} {Wave} {Functions}. {I}},
-  volume  = {23},
-  issn    = {0021-9606},
-  url     = {http://aip.scitation.org/doi/10.1063/1.1740588},
-  doi     = {10.1063/1.1740588},
-  number  = {10},
-  urldate = {2021-04-05},
-  journal = {The Journal of Chemical Physics},
-  author  = {Mulliken, R. S.},
-  month   = oct,
-  year    = {1955},
-  note    = {Publisher: American Institute of Physics},
-  pages   = {1833--1840}
-}
-
-@article{jakalian2000,
-  title      = {Fast, efficient generation of high-quality atomic charges. {AM1}-{BCC} model: {I}. {Method}},
-  volume     = {21},
-  copyright  = {Copyright © 2000 John Wiley \& Sons, Inc.},
-  issn       = {1096-987X},
-  shorttitle = {Fast, efficient generation of high-quality atomic charges. {AM1}-{BCC} model},
-  url        = {http://www.onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291096-987X%2820000130%2921%3A2%3C132%3A%3AAID-JCC5%3E3.0.CO%3B2-P},
-  doi        = {10.1002/(SICI)1096-987X(20000130)21:2<132::AID-JCC5>3.0.CO;2-P},
-  abstract   = {The AM1-BCC method quickly and efficiently generates high-quality atomic charges for use in condensed-phase simulations. The underlying features of the electron distribution including formal charge and delocalization are first captured by AM1 atomic charges for the individual molecule. Bond charge corrections (BCCs), which have been parameterized against the HF/6-31G* electrostatic potential (ESP) of a training set of compounds containing relevant functional groups, are then added using a formalism identical to the consensus BCI (bond charge increment) approach. As a proof of the concept, we fit BCCs simultaneously to 45 compounds including O-, N-, and S-containing functionalities, aromatics, and heteroaromatics, using only 41 BCC parameters. AM1-BCC yields charge sets of comparable quality to HF/6-31G* ESP-derived charges in a fraction of the time while reducing instabilities in the atomic charges compared to direct ESP-fit methods. We then apply the BCC parameters to a small “test set” consisting of aspirin, d-glucose, and eryodictyol; the AM1-BCC model again provides atomic charges of quality comparable with HF/6-31G* RESP charges, as judged by an increase of only 0.01 to 0.02 atomic units in the root-mean-square (RMS) error in ESP. Based on these encouraging results, we intend to parameterize the AM1-BCC model to provide a consistent charge model for any organic or biological molecule. © 2000 John Wiley \& Sons, Inc. J Comput Chem 21: 132–146, 2000},
-  language   = {en},
-  number     = {2},
-  urldate    = {2021-04-05},
-  journal    = {Journal of Computational Chemistry},
-  author     = {Jakalian, Araz and Bush, Bruce L. and Jack, David B. and Bayly, Christopher I.},
-  year       = {2000},
-  note       = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1002/\%28SICI\%291096-987X\%2820000130\%2921\%3A2\%3C132\%3A\%3AAID-JCC5\%3E3.0.CO\%3B2-P},
-  keywords   = {atomic charges, force field, bond charge correction, electrostatic potential, methodology},
-  pages      = {132--146}
-}
-
 @article{carlson1993,
[…]
