### 2022-03-28 #4100 @pdebuyl
Can you let me know when the plain conda install is working?

Apart from that, all is fine from my point of view, thanks for the updates!
