### 2022-04-04 a080cd3
add mamba back into CI
