### 2022-04-04 c16f5d7
add elementary error parsing in reading files
