+++ psiresp/qm.py
@@ -328,6 +328,8 @@ class BaseQMOptions(Model):
         with infile.open("rb") as f:
             content = f.read()
         data = qcel.util.deserialize(content, "msgpack")
+        if "model" not in data:  # there's an error
+            raise ValueError(data["error"])
         if data["model"]["basis"] == "":
             data["model"]["basis"] = None
         if "provenance" not in data:
