### 2022-04-04 d7d2919
Add elementary error parsing in reading files (#92)
