### 2022-04-15 08b1bd5
fix installation notes
