+++ devtools/conda-envs/environment.yaml
@@ -18,9 +18,9 @@ dependencies:
 
   # qcstack
   - postgresql >= 13.0
-  - qcelemental == 0.22.0
+  - qcelemental >= 0.22.0
   - qcengine
-  - qcfractal
+  - qcfractal >= 0.15
   - qcfractal-core
   - requests
 
