### 2022-04-15 bff35eb
Update installation notes (#94)
