### 2022-04-15 #4100 @lilyminium
@pdebuyl I added a new build with better pins as mentioned in https://github.com/openjournals/joss-reviews/issues/4100#issuecomment-1080704696, but that means that conda just ignores this new build in favour of the old ones without the pins, and therefore keeps pulling up the old version of qcfractal. I submitted a patch to update pins on the earlier builds https://github.com/conda-forge/conda-forge-repodata-patches-feedstock/pull/249 but that repository seems to move quite slowly, so it has not been reviewed yet.

In the mean time, I updated the installation notes on the PsiRESP to recommend people use mamba (which does not inexplicably favour old qcfractal) or specify their own pins: https://psiresp.readthedocs.io/en/latest/installation.html

That seems to be the only real solution for now, unfortunately.
