### 2022-04-19 #4100 @pdebuyl
`Solving environment: ...`

I'll provide an update but so far conda and mamba are slow.
