### 2022-04-20 #4100 @lilyminium
Thanks @pdebuyl. I didn't find mamba so bad, but `conda` can take forever -- so far as I can tell that's out of my control. I think adding the `--debug` flag can give you low level information on what it's doing.
