### 2022-04-22 #4100 @pdebuyl
Hi @lilyminium

Status: I could install with `mamba install -n tmpenv3.9 -c conda-forge psiresp==0.3.1 "qcfractal>0.15"`. I understand that psiresp is especially complex to install as it relies on quite elaborate packages that are not otherwise bundled.

I have only two requests before confirming the paper:
1. When running the examples, the user learns about the dependency on psi4 or mdanalysis, has to install and restart the kernel. Can you mention those dependencies visibly at the top of the example file?
2. The example "04_example_rest_minimal.ipynb" is not running because of missing input files.
