### 2022-04-23 #4100 @lilyminium
Hi @pdebuyl, I've done the first change! Unfortunately with regards to the second, that's because I added that example as a response to this review. `04_example_rest_minimal.ipynb` is not actually in version 0.3.1 of the repository, but the next release 0.4.0. That's why the input files are also missing, as they were created specifically for the example.

I looked into a bit and I can't patch a release I've already put on pip or conda -- however, I could make a new 0.3.2 that includes those examples as the only change from 0.3.1, and update the version in this review?
