### 2022-04-26 #4100 @pdebuyl
HI @lilyminium yes, a 0.3.2 release with only the minimal fixes to be able to run the examples properly would be totally ok.
