### 2022-04-29 #4100 @lilyminium
Great @pdebuyl, I've made a version 0.3.2 and put it on Zenodo here: https://zenodo.org/record/6501898 :)
