### 2022-05-04 #4100 @pdebuyl
There can be only one ISSN per bib entry. You have to pick one of "online" or "print" only.
