+++ paper/paper.bib
@@ -36,7 +36,7 @@
 @article{malde2011,
   title      = {An {Automated} {Force} {Field} {Topology} {Builder} ({ATB}) and {Repository}: {Version} 1.0},
   volume     = {7},
-  issn       = {1549-9618, 1549-9626},
+  issn       = {1549-9626},
   shorttitle = {An {Automated} {Force} {Field} {Topology} {Builder} ({ATB}) and {Repository}},
   url        = {https://pubs.acs.org/doi/10.1021/ct200196m},
   doi        = {10.1021/ct200196m},
@@ -57,7 +57,7 @@
 @article{cieplak1995,
   title      = {Application of the multimolecule and multiconformational {RESP} methodology to biopolymers: {Charge} derivation for {DNA}, {RNA}, and proteins},
   volume     = {16},
-  issn       = {0192-8651, 1096-987X},
+  issn       = {1096-987X},
   shorttitle = {Application of the multimolecule and multiconformational {RESP} methodology to biopolymers},
   url        = {http://doi.wiley.com/10.1002/jcc.540161106},
   doi        = {10.1002/jcc.540161106},
@@ -76,7 +76,7 @@
 @article{cornell1993,
   title    = {Application of {RESP} charges to calculate conformational energies, hydrogen bond energies, and free energies of solvation},
   volume   = {115},
-  issn     = {0002-7863, 1520-5126},
+  issn     = {1520-5126},
   url      = {http://pubs.acs.org/doi/abs/10.1021/ja00074a030},
   doi      = {10.1021/ja00074a030},
   abstract = {We apply a new restrained electrostatic potential fit charge model (two-stage RESP) to conformational analysis and the calculation of intermolecular interactions. Specifically, we study conformational energies in butane, methyl ethyl thioether, three simple alcohols, three simple amines, and 1,2-ethanediol as a function of charge model (two-stage RESP vs standard ESP) and 1-4 electrostatic scale factor. We demonstrate that the two-stage RESP model with a 1-4 electrostatic scale factor of —1/1.2 is a very good model, as evaluated by comparison with high-level ab initio calculations. For methanol and /V-methylacetamide interactions with TIP3P water, the two-stage RESP model leads to hydrogen bonds only slightly weaker than found with the standard ESP changes. In tests on DNA base pairs, the two-stage RESP model leads to hydrogen bonds which are {\textasciitilde} 1 kcal/mol weaker than those calculated with the standard ESP charges but closer in magnitude to the best current available ab initio calculations. Furthermore, the two-stage RESP charges, unlike the standard ESP charges, reproduce the result that Hoogsteen hydrogen bonding is stronger than Watson-Crick hydrogen bonding for adenine-thymine base pairs. The free energies of solvation of both methanol and f/ww-iV-methylacetamide were also calculated for the standard ESP and two-stage RESP models and both were in good agreement with experiment. We have combined the use of two-stage RESP charges with multiple conformational fitting—recently employed using standard ESP charges as described by Reynolds, et al. (J. Am. Chem. Soc. 1992, 114, 9075)—in studies of conformationally dependent dipole moments and energies of propylamine. We find that the combination of these approaches is synergistic in leading to useful charge distributions for molecular simulations. Two-stage RESP charges thus reproduce both intermolecular and intramolecular energies and structures quite well, making this charge model a critical advancement in the development of a general force field for modeling biological macromolecules and their ligands, both in the gas phase and in solution.},
@@ -96,7 +96,7 @@
 @article{bayly1993,
[…]
